fixing versions and clock tester helper
