+++ contracts/clock/Cargo.toml
@@ -26,7 +26,6 @@ cw2 = { workspace = true }
 serde = { workspace = true }
 thiserror = { workspace = true }
 neutron-sdk      = { workspace = true }
-cw-multi-test = { workspace = true }
 
 [dev-dependencies]
 cw-multi-test = { workspace = true }
