+++ contracts/swap-holder/src/suite_tests/suite.rs
@@ -1,8 +1,9 @@
-use crate::msg::{ContractState, ExecuteMsg, InstantiateMsg, QueryMsg};
-use cosmwasm_std::{Addr, Coin, Uint128};
-use covenant_clock::test_helpers::helpers::{
-    mock_clock_deps_contract, mock_clock_instantiate_message,
+use crate::{
+    msg::{ContractState, ExecuteMsg, InstantiateMsg, QueryMsg},
+    suite_tests::mock_clock_deps_contract,
 };
+use cosmwasm_std::{Addr, Coin, Uint128};
+use covenant_clock::test_helpers::helpers::mock_clock_instantiate_message;
 use covenant_utils::{
     CovenantPartiesConfig, CovenantParty, CovenantTerms, ReceiverConfig, SwapCovenantTerms,
 };
