+++ contracts/two-party-pol-holder/src/suite_tests/suite.rs
@@ -1,13 +1,14 @@
 use std::{collections::BTreeMap, str::FromStr};
 
-use crate::msg::{
-    ContractState, CovenantType, ExecuteMsg, InstantiateMsg, QueryMsg, RagequitConfig,
-    TwoPartyPolCovenantConfig, TwoPartyPolCovenantParty,
+use crate::{
+    msg::{
+        ContractState, CovenantType, ExecuteMsg, InstantiateMsg, QueryMsg, RagequitConfig,
+        TwoPartyPolCovenantConfig, TwoPartyPolCovenantParty,
+    },
+    suite_tests::mock_clock_deps_contract,
 };
 use cosmwasm_std::{Addr, BlockInfo, Coin, Decimal, Timestamp, Uint128};
-use covenant_clock::test_helpers::helpers::{
-    mock_clock_deps_contract, mock_clock_instantiate_message,
-};
+use covenant_clock::test_helpers::helpers::mock_clock_instantiate_message;
 use covenant_utils::{SplitConfig, SplitType};
 use cw_multi_test::{App, AppResponse, Executor, SudoMsg};
 use cw_utils::Expiration;
