+++ optimize.sh
@@ -3,12 +3,12 @@ if [[ $(uname -m) =~ "arm64" ]]; then \
     docker run --rm -v "$(pwd)":/code \
         --mount type=volume,source="$(basename "$(pwd)")_cache",target=/code/target \
         --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
-        cosmwasm/workspace-optimizer-arm64:0.12.4
+        cosmwasm/workspace-optimizer-arm64:0.14.0
 
 else
     docker run --rm -v "$(pwd)":/code \
         --mount type=volume,source="$(basename "$(pwd)")_cache",target=/code/target \
         --mount type=volume,source=registry_cache,target=/usr/local/cargo/registry \
         --platform linux/amd64 \
-        cosmwasm/workspace-optimizer:0.12.4
+        cosmwasm/workspace-optimizer:0.14.0
 fi
