+++ packages/covenant-macros/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "covenant-macros"
-version = "0.0.1"
-edition = "2021"
+version = { workspace = true }
+edition =  { workspace = true }
 authors = ["ekez <ekez@withoutdoing.com>, benskey bekauz@protonmail.com"]
 description = "A package for deriving the covenant interfaces."
 license = "BSD-3"
